Refactored src and main
