+++ CS320_djhake2/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>CS320_djhake2</name>
+	<name>CS320_JKSOrders</name>
 	<comment></comment>
 	<projects>
 	</projects>
