+++ CS320_djhake2/junit/edu/ycp/cs320/djhake2/controller/GuessingGameControllerTest.java
@@ -5,8 +5,8 @@ import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
 
-import edu.ycp.cs320.djhake2.controller.GuessingGameController;
-import edu.ycp.cs320.djhake2.model.GuessingGame;
+import edu.ycp.cs320.JKSOrders.controller.GuessingGameController;
+import edu.ycp.cs320.JKSOrders.model.GuessingGame;
 
 public class GuessingGameControllerTest {
 	private GuessingGame model;
