+++ CS320_djhake2/junit/edu/ycp/cs320/djhake2/model/GuessingGameTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.*;
 import org.junit.Before;
 import org.junit.Test;
 
-import edu.ycp.cs320.djhake2.model.GuessingGame;
+import edu.ycp.cs320.JKSOrders.model.GuessingGame;
 
 public class GuessingGameTest {
 	private GuessingGame model;
