+++ CS320_djhake2/src/edu/ycp/cs320/JKSOrders/servlet/ajax/NextGuessAjaxServlet.java
@@ -1,4 +1,4 @@
-package edu.ycp.cs320.djhake2.servlet.ajax;
+package edu.ycp.cs320.JKSOrders.servlet.ajax;
 
 import java.io.IOException;
 
@@ -7,8 +7,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import edu.ycp.cs320.djhake2.controller.GuessingGameController;
-import edu.ycp.cs320.djhake2.model.GuessingGame;
+import edu.ycp.cs320.JKSOrders.controller.GuessingGameController;
+import edu.ycp.cs320.JKSOrders.model.GuessingGame;
 
 public class NextGuessAjaxServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
